add news generator and fixed some bugs

+++ backend/src/main/java/com/lerhyd/dngame/dao/AchievementDao.java
@@ -12,9 +12,9 @@ import java.util.Optional;
 @Repository
 public interface AchievementDao extends JpaRepository<Achievement, String> {
 
-    List<Achievement> findAchievementsByKiras(long id);
+    List<Achievement> findAchievementsByKiras(int id);
 
-    List<Achievement> findAchievementsByAgents(long id);
+    List<Achievement> findAchievementsByAgents(int id);
 
     Optional<Achievement> findById(String id);
 }
