+++ backend/src/main/java/com/lerhyd/dngame/dao/ActionDao.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActionDao extends JpaRepository<Action, Long> {
+public interface ActionDao extends JpaRepository<Action, Integer> {
 
-    Action findById(long id);
+    Action findById(int id);
 
 }
