+++ backend/src/main/java/com/lerhyd/dngame/dao/ActionPlaceDao.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActionPlaceDao extends JpaRepository<ActionPlace, Long> {
+public interface ActionPlaceDao extends JpaRepository<ActionPlace, Integer> {
 
-    ActionPlace findById(long id);
+    ActionPlace findById(int id);
 
 }
