+++ backend/src/main/java/com/lerhyd/dngame/dao/RankDao.java
@@ -7,15 +7,15 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RankDao extends JpaRepository<Rank, Long> {
+public interface RankDao extends JpaRepository<Rank, Integer> {
 
     @Query("select r.rank from Rank r left join r.agents a left join a.user u where u.id = :userId")
-    String findRankByUserIdIfIsAgents(@Param("userId") Long userId);
+    String findRankByUserIdIfIsAgents(@Param("userId") int userId);
 
     @Query("select r.rank from Rank r left join r.kiras k left join k.user u where u.id = :userId")
-    String findRankByUserIdIfIsKiras(@Param("userId") Long userId);
+    String findRankByUserIdIfIsKiras(@Param("userId") int userId);
 
-    Rank findById(long id);
+    Rank findById(int id);
 
     @Query("select r from Rank r where r.points = :points and r.kira=:isKira")
     Rank findByPoints(@Param("points") int points,@Param("isKira") boolean isKira);
