+++ backend/src/main/java/com/lerhyd/dngame/request/NewsReq.java
@@ -8,13 +8,13 @@ public class NewsReq {
     private boolean die;
     private String desc;
     private String pubDate;
-    private long actionId;
-    private long actionPlaceId;
-    private long commonRegionId;
-    private long distRegionId;
-    private long agentId;
-    private long kiraId;
-    private long victimId;
-    private long guiltyPersonId;
+    private int actionId;
+    private int actionPlaceId;
+    private int commonRegionId;
+    private int distRegionId;
+    private int agentId;
+    private int kiraId;
+    private int victimId;
+    private int guiltyPersonId;
 
 }
